fix: Feature of services-hdfs not handled correctly
