+++ .github/workflows/service_test_azblob.yml
@@ -27,7 +27,7 @@ jobs:
 
       - name: Test
         shell: bash
-        run: cargo test azblob --features compress,retry -- --nocapture
+        run: cargo test azblob --features compress,retry,testing -- --nocapture
         env:
           RUST_BACKTRACE: full
           RUST_LOG: debug
