+++ .github/workflows/service_test_hdfs.yml
@@ -33,7 +33,7 @@ jobs:
 
       - name: Test
         shell: bash
-        run: cargo test hdfs --features compress,retry,services-hdfs -- --nocapture
+        run: cargo test hdfs --features compress,retry,testing,services-hdfs -- --nocapture
         env:
           RUST_BACKTRACE: full
           RUST_LOG: debug
@@ -69,7 +69,7 @@ jobs:
 
       - name: Test
         shell: bash
-        run: cargo test hdfs --features compress,retry,services-hdfs -- --nocapture
+        run: cargo test hdfs --features compress,retry,testing,services-hdfs -- --nocapture
         env:
           RUST_BACKTRACE: full
           RUST_LOG: debug
