+++ .github/workflows/service_test_memory.yml
@@ -18,7 +18,7 @@ jobs:
       - uses: actions/checkout@v3
       - name: Test
         shell: bash
-        run: cargo test memory --features compress,retry -- --nocapture
+        run: cargo test memory --features compress,retry,testing -- --nocapture
         env:
           RUST_BACKTRACE: full
           RUST_LOG: debug
