+++ .github/workflows/service_test_s3.yml
@@ -13,7 +13,7 @@ jobs:
       - uses: actions/checkout@v3
       - name: Test
         shell: bash
-        run: cargo test s3 --features compress,retry -- --nocapture
+        run: cargo test s3 --features compress,retry,testing -- --nocapture
         env:
           RUST_BACKTRACE: full
           RUST_LOG: debug
@@ -30,7 +30,7 @@ jobs:
       - uses: actions/checkout@v3
       - name: Test
         shell: bash
-        run: cargo test s3 --features compress,retry -- --nocapture
+        run: cargo test s3 --features compress,retry,testing -- --nocapture
         env:
           RUST_BACKTRACE: full
           RUST_LOG: debug
@@ -68,7 +68,7 @@ jobs:
 
       - name: Test
         shell: bash
-        run: cargo test s3 --features compress,retry -- --nocapture
+        run: cargo test s3 --features compress,retry,testing -- --nocapture
         env:
           RUST_BACKTRACE: full
           RUST_LOG: debug
@@ -103,7 +103,7 @@ jobs:
 
       - name: Test
         shell: bash
-        run: cargo test s3 --features compress,retry -- --nocapture
+        run: cargo test s3 --features compress,retry,testing -- --nocapture
         env:
           RUST_BACKTRACE: full
           RUST_LOG: debug
