+++ Cargo.toml
@@ -15,10 +15,8 @@ all-features = true
 [features]
 compress = ["async-compression"]
 retry = ["backon"]
-services-hdfs = ["hdrs", "opendal_test/services-hdfs"]
-
-[workspace]
-members = ["opendal_test"]
+services-hdfs = ["hdrs"]
+testing = ["dotenv", "uuid"]
 
 [lib]
 bench = false
@@ -35,8 +33,8 @@ name = "io"
 anyhow = "1.0.56"
 async-compat = "0.2.1"
 async-compression = { version = "0.3.12", features = [
-  "futures-io",
-  "all-algorithms",
+    "futures-io",
+    "all-algorithms",
 ], optional = true }
 async-trait = "0.1.53"
 backon = { version = "0.0.2", optional = true }
@@ -62,22 +60,23 @@ serde = { version = "1.0.136", features = ["derive"] }
 thiserror = "1.0.30"
 time = "0.3.9"
 tokio = { version = "1.17.0", features = ["full"] }
+dotenv = { version = "0.15.0", optional = true }
+uuid = { version = "1.0.0", optional = true, features = ["serde", "v4"] }
 
 [dev-dependencies]
 anyhow = "1.0.56"
 criterion = { version = "0.3.5", features = [
-  "async",
-  "async_tokio",
-  "html_reports",
+    "async",
+    "async_tokio",
+    "html_reports",
 ] }
-dotenv = "0.15.0"
 env_logger = "0.9.0"
 itertools = "0.10.3"
 num-traits = "0.2.14"
-opendal_test = { path = "./opendal_test" }
 paste = "1.0.7"
 rand = "0.8.5"
 serde_json = "1.0.79"
 sha2 = "0.10.2"
 size = "0.1.2"
 uuid = { version = "1.0.0", features = ["serde", "v4"] }
+cfg-if = "1.0.0"
