--- opendal_test/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-authors = ["Databend Authors <opensource@datafuselabs.com>"]
-categories = ["filesystem"]
-description = "Open Data Access Layer that connect the whole world together."
-edition = "2021"
-keywords = ["storage", "data", "s3"]
-license = "Apache-2.0"
-name = "opendal_test"
-publish = false
-repository = "https://github.com/datafuselabs/opendal"
-version = "0.0.0"
-
-[features]
-services-hdfs = ["opendal/services-hdfs"]
-
-[dependencies]
-dotenv = "0.15.0"
-opendal = { path = ".." }
-uuid = { version = "1.0.0", features = ["serde", "v4"] }
