+++ src/services/hdfs/mod.rs
@@ -92,3 +92,7 @@ pub use backend::Builder;
 
 mod error;
 mod object_stream;
+
+#[doc(hidden)]
+#[cfg(feature = "testing")]
+pub mod tests;
