+++ src/services/s3/mod.rs
@@ -71,3 +71,7 @@ pub use backend::Backend;
 pub use backend::Builder;
 
 mod object_stream;
+
+#[doc(hidden)]
+#[cfg(feature = "testing")]
+pub mod tests;
